requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
python-multipart==0.0.12
pdfminer.six==20231228
bcrypt==4.0.1
